requirements.txt
azure-functions==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pyodbc==5.1.0
requests==2.32.3
urllib3==2.2.2
mysql-connector-python==8.0.33 